refactor: enum - 파일명 변경
