+++ services/apply.js
@@ -2,7 +2,7 @@ const applyDao = require('../dao/apply');
 
 const { rowSplit, toBoolean } = require('../utils/query');
 const { customError } = require('../utils/errors/customError');
-const { applyEnum } = require('../utils/variables/enums');
+const { applyEnum } = require('../utils/variables/enum');
 const broadcast = require('../events/broadcast');
 
 const User = require('../models/user');
