+++ utils/query.js
@@ -57,7 +57,7 @@ const cutId = (rows) => {
   return rows;
 };
 
-const customSorting = (sigungu, rows) => {
+const lengthSorting = (sigungu, rows) => {
   let [targetRows, otherRows] = [[], []];
 
   rows.forEach((row, idx) => {
@@ -71,10 +71,25 @@ const customSorting = (sigungu, rows) => {
   return targetRows.concat(otherRows);
 };
 
+const tokenDivision = (memberRows) => {
+  const [fcm_token, apns_token] = [[], []];
+
+  memberRows.forEach((v) => {
+    if (v.device === 'ios') {
+      apns_token.push(v.push_token);
+    } else {
+      fcm_token.push(v.push_token);
+    }
+  });
+
+  return [apns_token, fcm_token];
+};
+
 module.exports = {
   rowSplit,
   toBoolean,
   locationMerge,
   cutId,
-  customSorting,
+  lengthSorting,
+  tokenDivision,
 };
